requirements.txt
fastapi
uvicorn[standard]
mistralai
pdfplumber
jsonschema
python-dotenv
pydantic>=2
